# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Energy Monitor API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/config/host/{id}": {
            "delete": {
                "tags": [
                    "Hosts"
                ],
                "summary": "Delete a host in Emca Monitor",
                "description": "Deletes an existing host.",
                "operationId": "deleteHost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the host to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Host deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Host deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Host not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Host not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
        "/api/config/hostgroup": {
            "get": {
                "tags": [
                    "HostGroups"
                ],
                "summary": "Get all Host Groups or single hostGroup in Emca Monitor",
                "description": "Retrieves a list of all Host Groups or a specific hostGroup by ID or name.",
                "operationId": "getAllhostGroups",
                "parameters": [
                    {
                        "name": "identifier",
                        "in": "query",
                        "description": "ID or name of the hostGroup (optional)",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Host Groups",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "HostGroups"
                ],
                "summary": "Create a new HostGroup in Emca Monitor",
                "description": "Creates a new HostGroup with the provided parameters.",
                "operationId": "createHostGroups",
                "requestBody": {
                    "description": "command details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "hostgroup_name",
                                    "file_id"
                                ],
                                "properties": {
                                    "hostgroup_name": {
                                        "type": "string",
                                        "example": "host_group"
                                    },
                                    "file_id": {
                                        "type": "string",
                                        "example": "/opt/naemon/etc/naemon/conf.d/localhost.cfg"
                                    },
                                    "action_url": {
                                        "type": "string",
                                        "example": "notify-host-by-email"
                                    },
                                    "alias": {
                                        "type": "string",
                                        "example": "..alias"
                                    },
                                    "hostgroup_members": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "hostGroup_member_id"
                                        }
                                    },
                                    "members": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "member_id"
                                        }
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "notes_url": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "register": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Host Group created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Host Group created successfully"
                                        },
                                        "hostGroup": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "Host Group description": [
                                                    "The Host Group description field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "basic"
            }
        }
    }
}